# openapi2.json
{
  "openapi": "3.0.0",
  "info": {
    "version": "1.0.0",
    "title": "Hobotti API",
    "description": "A basic backend API for the hobby/likemided community finder mobile app Hobotti. <br> Please contact the developers to gain access to the API key needed to use this API.",
    "contact": {
      "email": "santtu-taskinen@outlook.com"
    }
  },
...
  "paths": {
...
    "/events/{search}": {
      "get": {
        "security": [
          {
            "APIAccess": [],
            "HobottiAccess": []
          }
        ],
        "summary": "Get event data from Linked Events",
        "tags": [
          "Events"
        ],
        "parameters": [
          {
            "name": "search",
            "in": "path",
            "description": "Hobby you want find events for.",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "page",
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Events Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/eventlist"
                }
              }
            }
          },
          "404": {
            "description": "No events found for provided search term.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/messageResponse"
                }
              }
            }
          },
          "463": {
            "description": "No API Key",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/messageResponse"
                }
              }
            }
          },
          "464": {
            "description": "Invalid API key",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/messageResponse"
                }
              }
            }
          },
          "473": {
            "description": "No access token provided.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/messageResponse"
                }
              }
            }
          },
          "500": {
            "description": "Something went wrong",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/messageResponse"
                }
              }
            }
          },
          "565": {
            "description": "Something went wrong with the API key verification.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/messageResponse"
                }
              }
            }
          },
          "570": {
            "description": "Something went wrong while trying to authenticate token.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/messageResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/logs/events/{username}": {
      "get": {
        "security": [
          {
            "APIAccess": [],
            "HobottiAccess": []
          }
        ],
        "summary": "Get the last twenty chat events (currently only joining and leaving chatrooms) of a specific user.",
        "tags": [
          "Logs"
        ],
        "parameters": [
          {
            "name": "username",
            "in": "path",
            "description": "The name of the user whose movements you want to follow.",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Messages found successfully",
            "content": {
              "application/array": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/eventLogItem"
                  }
                }
              }
            }
          },
          "404": {
            "description": "No messages were found in the logs for the given chatroom.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/messageResponse"
                }
              }
            }
          },
          "463": {
            "description": "No API Key",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/messageResponse"
                }
              }
            }
          },
          "464": {
            "description": "Invalid API key",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/messageResponse"
                }
              }
            }
          },
          "473": {
            "description": "No access token provided.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/messageResponse"
                }
              }
            }
          },
          "500": {
            "description": "Something went wrong serverside while searching the logs.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/messageResponse"
                }
              }
            }
          },
          "565": {
            "description": "Something went wrong with the API key verification.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/messageResponse"
                }
              }
            }
          },
          "570": {
            "description": "Something went wrong while trying to authenticate token.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/messageResponse"
                }
              }
            }
          }
        }
      }
    }
  },
  "servers": [
    {
      "url": "https://hobotti-backend-testing.herokuapp.com/api/"
    },
    {
      "url": "http://hobotti-backend-testing.herokuapp.com/api/"
    },
    {
      "url": "http://localhost:3000/api/"
    }
  ],
  "components": {
    "securitySchemes": {
      "APIAccess": {
        "type": "apiKey",
        "in": "header",
        "name": "hobotti-api-key"
      },
      "HobottiAccess": {
        "type": "apiKey",
        "in": "header",
        "name": "hobotti-access-token"
      }
    },
    "schemas": {
...
      "eventLogItem": {
        "properties": {
          "event": {
            "type": "string"
          },
          "user": {
            "type": "string"
          },
          "date": {
            "type": "string"
          }
        }
      },
...
      "eventlist": {
        "properties": {
          "count": {
            "type": "integer"
          },
          "fetchDate": {
            "type": "string"
          },
          "events": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/event"
            }
          }
        }
      },
      "event": {
        "properties": {
          "id": {
            "type": "string"
          },
          "event_type": {
            "$ref": "#/components/schemas/multilingual"
          },
          "name": {
            "$ref": "#/components/schemas/multilingual"
          },
          "provider": {
            "$ref": "#/components/schemas/multilingual"
          },
          "s_desc": {
            "$ref": "#/components/schemas/multilingual"
          },
          "desc": {
            "$ref": "#/components/schemas/multilingual"
          },
          "info": {
            "type": "string"
          },
          "start_time": {
            "type": "string"
          },
          "end_time": {
            "type": "string"
          },
          "price": {
            "$ref": "#/components/schemas/multilingual"
          },
          "images": {
            "type": "array",
            "items": {
              "type": "object"
            }
          }
        }
      },
      "multilingual": {
        "properties": {
          "sv": {
            "type": "string"
          },
          "en": {
            "type": "string"
          },
          "fi": {
            "type": "string"
          }
        }
      },
...
      "messageResponse": {
        "properties": {
          "message": {
            "type": "string"
          }
        }
      },
...
    }
  },
  "security": [
    {
      "APIAccess": [],
      "HobottiAccess": []
    }
  ]
}